improve main.py
init と error を別ファイルに書き出した。
