+++ python/src/main.py
@@ -1,12 +1,8 @@
-from logging import getLogger, StreamHandler, DEBUG, Formatter, FileHandler
 import feedparser
-import os
 import db.database as db
-import dotenv
-import sys
 import interfaces.twitter as twitter
 import interfaces.mattermost as mattermost
-import slackweb
+import init
 
 #########################
 # cronで回す時は相対パスを絶対パスにしないといけない
@@ -16,19 +12,17 @@ import slackweb
 # endpointの設定
 ENDPOINT = "http://rais.skr.u-ryukyu.ac.jp/dc/?feed=rss2"
 
-ENVIRONMENT_FILE = "./../.env"
+
 
 ##########
 LIVE_SERVER = 0
 TEST_SERVER = 1
 DRY_RUN = 2
-
-
 ##########
 
 
 def main():
-    status, logger, err = init()
+    status, logger, err = init.init()
 
     # ENDPOINTにアクセスしrssを取得
     logger.info("Access to endpoint")
@@ -71,82 +65,6 @@ def main():
     cur.close()
     database.conn.close()
     logger.info("Successful completion")
-
-
-def init_logger(modname=__name__):
-    error_flg = False
-    try:
-        log_folder = os.environ.get('LOG_PATH')
-    except KeyError:
-        log_folder = "./../sedona.log"
-        error_flg = True
-    logger = getLogger(modname)
-    logger.setLevel(DEBUG)
-    sh = StreamHandler()
-    sh.setLevel(DEBUG)
-    formatter = Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    sh.setFormatter(formatter)
-    logger.addHandler(sh)
-
-    fh = FileHandler(log_folder)  # fh = file handler
-    fh.setLevel(DEBUG)
-    fh_formatter = Formatter('%(asctime)s - %(filename)s - %(name)s - %(lineno)d - %(levelname)s - %(message)s')
-    fh.setFormatter(fh_formatter)
-    logger.addHandler(fh)
-    if error_flg:
-        logger.info("Can not read LOG_PATH. \nPlease check environment variable.")
-    return logger
-
-
-# logging の設定と引数の処理を行います。 戻り値にスクリプト実行時の引数と logger を返します。.
-def init():
-    dotenv.load_dotenv(ENVIRONMENT_FILE)
-    logger = init_logger(modname="debug")
-    # loggingの設定
-    logger.info("sedona start")
-    # errorの設定
-    error = Error(logger)
-    # 引数の処理
-    status = DRY_RUN  # とりあえず安全な値を初期値にする
-    try:
-        args = sys.argv
-        if len(args) != 2:
-            logger.error("Arguments Error")
-            logger.error("Example python main.py dry-run")
-            error.send("Arguments Error")
-            exit(1)
-
-        if args[1] == "live-server":
-            logger.info("THIS IS LIVE-SERVER")
-            status = LIVE_SERVER
-        elif args[1] == "test-server":
-            logger.info("THIS IS TEST-SERVER")
-            status = TEST_SERVER
-        elif args[1] == "dry-run":
-            logger.info("THIS IS DRY-RUN")
-
-    except KeyError as e:
-        # .env関係のエラー
-        logger.error("Environment variable error")
-        logger.error(e)
-        error.send(e)
-        exit(1)
-    return status, logger, error
-
-
-class Error:
-    def __init__(self, logger):
-        try:
-            self.address = os.environ.get('ADDRESS')
-            self.logger = logger
-            pass
-        except KeyError as e:
-            self.logger(e)
-            exit(1)
-
-    def send(self, message):
-        error_chat_room = slackweb.Slack(url=self.address)
-        error_chat_room.notify(text="From sedona: " + str(message))
 
 
 if __name__ == '__main__':
